formatting and move file
