--- src/Illuminate/Log/LoggedExceptionCollection.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Illuminate\Log;
-
-use Illuminate\Support\Collection;
-
-class LoggedExceptionCollection extends Collection
-{
-    //
-}
